requirements.txt
fastapi[standard]
sqlalchemy
python-dotenv
jinja2
pyjwt
passlib[bcrypt]